requirements.txt
pandas
numpy
python-dotenv
pymongo
certifi
pymongo[srv]
scikit-learn
# -e .
